MAINT: Add a Python wrapper for git cloning
